backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-dotenv==1.0.0
PyJWT==2.8.0
django-filter==23.5
Pillow==10.2.0  # Para manejo de imágenes
gunicorn==21.2.0  # Para servidor en producción
